src/requirements.txt
boto3==1.38.0
botocore==1.38.0
llama_index==0.12.33
pandas==2.2.3
pydantic==2.11.3
python-dotenv==1.1.0
python-telegram-bot==22.0
tiktoken==0.9.0
